+++ bin/stop-server.sh
@@ -9,7 +9,7 @@ stop_server () {
     if [ -e "$1" ]; then
         PID=$(cat "$1");
         echo PID="$PID";
-        # Killing juts the PID is not enough sometimes, nodejs server still lingers on (at least on Windows)
+        # Killing just the PID is not enough sometimes, nodejs server still lingers on (at least on Windows)
         if [ -n "$PID" ]; then
             kill "$PID" || true;
         fi;