+++ tests/Functional/ValidatorTest.php
@@ -1,24 +1,26 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace SMTPValidateEmail\Tests\Functional;
 
+use SMTPValidateEmail\Exceptions\Exception;
 use SMTPValidateEmail\Exceptions\NoConnection;
 use SMTPValidateEmail\Validator;
 use SMTPValidateEmail\Tests\TestCase;
+use function array_pop;
 
 /**
  * Functional tests for Validator class.
  */
 class ValidatorTest extends TestCase
 {
-    const CONNECT_PORT = 1025;
+    public const CONNECT_PORT = 1025;
 
-    public function testNoConnIsValid()
+    public function testNoConnIsValid(): void
     {
         // Testing no connection, so using a non-existent host on .localhost
         $uniq    = uniqid();
         $host    = 'localhost';
-        $email   = 'test@' . $uniq . '.' . $host;
+        $email   = 'test@' . $uniq . '.' . $host . '.'; // Malformed on purpose, so MX resolution bombs very early
         $timeout = 1;
 
         // Default should be false, and there's no smtp on localhost:25
@@ -29,7 +31,7 @@ class ValidatorTest extends TestCase
         $this->assertFalse($results[$email]);
 
         $log       = $inst->getLog();
-        $last_line = \array_pop($log);
+        $last_line = array_pop($log);
         $needle    = 'Unable to connect. Exception caught: Cannot open a connection to remote host';
         $this->assertStringContainsString($needle, $last_line);
 
@@ -40,14 +42,38 @@ class ValidatorTest extends TestCase
         $this->assertTrue($results[$email]);
     }
 
+    public function testBindAddressWorking(): void
+    {
+        if (!$this->isSmtpServerRunning()) {
+            $this->markTestSkipped('smtp server not running.');
+        }
+
+        $this->disableJim(); // Prevent random misbehavior
+
+        $email = 'test@localhost';
+        $inst = new Validator($email, 'hello@localhost');
+        $inst->setConnectTimeout(1);
+        $inst->setConnectPort(self::CONNECT_PORT);
+
+        $bindPort = self::CONNECT_PORT + 1000;
+        $bindAddress = '127.0.0.1:' . $bindPort;
+
+        $inst->setBindAddress($bindAddress);
+        $inst->validate();
+
+        // Assert smtp server's log file shows communication happened on configured $bindAddress
+        $logFileContents = file_get_contents('/tmp/mailhog.log');
+        $this->assertStringContainsString('[SMTP ' . $bindAddress . ']', $logFileContents);
+    }
+
     /**
      * Requires having the smtp-server running locally, configured to
      * reject any sender or only allow a certain sender (not much difference really).
      */
-    public function testNoCommIsValidWithLocalSmtpRejectingOurSender()
+    public function testNoCommIsValidWithLocalSmtpRejectingOurSender(): void
     {
         if (!$this->isSmtpServerRunning()) {
-            $this->markTestSkipped('SMTP server not running.');
+            $this->markTestSkipped('smtp server not running.');
         }
 
         // Re-configure mailhog to reject any sender we give it.
@@ -71,7 +97,7 @@ class ValidatorTest extends TestCase
         $this->disableJim();
     }
 
-    public function testValidSenderWithLocalSmtp()
+    public function testValidSenderWithLocalSmtp(): void
     {
         if (!$this->isSmtpServerRunning()) {
             $this->markTestSkipped('smtp server not running.');
@@ -87,7 +113,7 @@ class ValidatorTest extends TestCase
         $this->assertTrue($results[$email]);
     }
 
-    public function testCatchAllConsideredValid()
+    public function testCatchAllConsideredValid(): void
     {
         if (!$this->isSmtpServerRunning()) {
             $this->markTestSkipped('smtp server not running.');
@@ -109,7 +135,7 @@ class ValidatorTest extends TestCase
         }
     }
 
-    public function testCatchAllConsideredInvalid()
+    public function testCatchAllConsideredInvalid(): void
     {
         if (!$this->isSmtpServerRunning()) {
             $this->markTestSkipped('smtp server not running.');
@@ -129,11 +155,11 @@ class ValidatorTest extends TestCase
         $this->assertFalse($results[$email]);
     }
 
-    public function testRandomDomainWithNoMxRecords()
+    public function testRandomDomainWithNoMxRecords(): void
     {
         $uniq  = uniqid();
         $host  = $uniq . '.com';
-        $email = 'test@' . $uniq . '.' . $host;
+        $email = 'test@' . $uniq . '.' . $host . '.'; // Dot in the end to bomb MX query very quickly
 
         // Default no_conn_is_valid should be false
         $inst = new Validator($email, 'hello@localhost');
@@ -142,25 +168,35 @@ class ValidatorTest extends TestCase
         $this->assertFalse($results[$email]);
 
         $log       = $inst->getLog();
-        $last_line = \array_pop($log);
+        $last_line = array_pop($log);
         $needle    = 'Unable to connect. Exception caught: Cannot open a connection to remote host';
         $this->assertStringContainsString($needle, $last_line);
     }
 
-    public function testIssue35()
+    public function testIssue35(): void
     {
-        $email   = 'blabla@blablabla.bla';
-        $inst    = new Validator($email, 'hello@localhost');
+        // Email on a non-existing domain taking forever and/or "crashing".
+
+        // Prevent random misbehavior
+        if ($this->isSmtpServerRunning()) {
+            $this->disableJim();
+        }
+
+        // Note: the trailing period on a non-existing domain name makes the MX
+        // query fail fast as opposed to the DNS resolver doing relative-to-absolute-fqdn appends/checks
+        // with common suffixes etc. (which all fail in the end anyway).
+        $email = 'blabla@blablabla.bla';
+        $inst = new Validator($email, 'hello@localhost');
         $results = $inst->validate();
         $this->assertFalse($results[$email]);
 
-        $log       = $inst->getLog();
-        $last_line = \array_pop($log);
-        $needle    = 'Unable to connect. Exception caught: Cannot open a connection to remote host';
+        $log = $inst->getLog();
+        $last_line = array_pop($log);
+        $needle = 'Unable to connect. Exception caught: Cannot open a connection to remote host';
         $this->assertStringContainsString($needle, $last_line);
     }
 
-    public function testNoopsSentByDefault()
+    public function testNoopsSentByDefault(): void
     {
         if (!$this->isSmtpServerRunning()) {
             $this->markTestSkipped('smtp server not running.');
@@ -173,10 +209,10 @@ class ValidatorTest extends TestCase
         $inst->setConnectPort(self::CONNECT_PORT);
         $inst->validate();
         $log = $inst->getLog();
-        $this->assertRegexp('/NOOP/', implode('', $log));
+        $this->assertMatchesRegularExpression('/NOOP/', implode('', $log));
     }
 
-    public function testNoopsDisabled()
+    public function testNoopsDisabled(): void
     {
         if (!$this->isSmtpServerRunning()) {
             $this->markTestSkipped('smtp server not running.');
@@ -190,10 +226,10 @@ class ValidatorTest extends TestCase
         $inst->sendNoops(false);
         $inst->validate();
         $log = $inst->getLog();
-        $this->assertNotRegexp('/NOOP/', implode('', $log));
+        $this->assertDoesNotMatchRegularExpression('/NOOP/', implode('', $log));
     }
 
-    public function testRejectedConnection()
+    public function testRejectedConnection(): void
     {
         if (!$this->isSmtpServerRunning()) {
             $this->markTestSkipped('smtp server not running.');
@@ -204,7 +240,7 @@ class ValidatorTest extends TestCase
         // Configure Jim how we need it
         $this->makeSmtpRejectConnections();
 
-        $test = function () {
+        $test = static function () {
             $email = 'chaos@localhost';
             $inst  = new Validator($email, 'alice@localhost');
             $inst->setConnectTimeout(1);
@@ -218,19 +254,19 @@ class ValidatorTest extends TestCase
         $this->disableJim();
     }
 
-    public function testDisconnections()
+    public function testDisconnections(): void
     {
         if (!$this->isSmtpServerRunning()) {
             $this->markTestSkipped('smtp server not running.');
         }
 
         // Hopefully we hit at least one of our exceptions...
-        $this->expectException(\SMTPValidateEmail\Exceptions\Exception::class);
+        $this->expectException(Exception::class);
 
         // Configure Jim how we need it
         $this->makeSmtpRandomlyDisconnect();
 
-        $test = function () {
+        $test = static function () {
             $email = 'disconnector@localhost';
             $inst  = new Validator($email, 'checker@localhost');
             $inst->setConnectTimeout(1);